Fix issue #32 Parent dirs in elfeed-score-rule-stats-file are not
Setting `elfeed-score-rule-stats-file` to a path with intermediate
directories that don't exist will error-out. This fix will at
least attempt to create them.

Also, update the CI workflow & generally prep for the merge
(new versions in docs &c)

As an aside, the regression test was originally written using
`file-name-concat' which was only introduced in Emacs 28.1
(we support 26.1). It now uses a purpose-build Emacs Lisp
replacement.